@@ -19,8 +19,9 @@ namespace TeduCoreApp.Areas.Admin.Controllers
     public class BillController : BaseController
     {
         private readonly IBillService _billService;
-        private readonly IHostingEnvironment _hostingEnvironment;
-        public BillController(IBillService billService, IHostingEnvironment hostingEnvironment)
+        private readonly IWebHostEnvironment _hostingEnvironment;
+
+        public BillController(IBillService billService, IWebHostEnvironment hostingEnvironment)
         {
             _billService = billService;
             _hostingEnvironment = hostingEnvironment;
@@ -38,6 +39,7 @@ namespace TeduCoreApp.Areas.Admin.Controllers
 
             return new OkObjectResult(model);
         }
+
         [HttpGet]
         public IActionResult UpdateStatus(int billId, BillStatus status)
         {
@@ -45,12 +47,14 @@ namespace TeduCoreApp.Areas.Admin.Controllers
 
             return new OkResult();
         }
+
         [HttpGet]
         public IActionResult GetAllPaging(string startDate, string endDate, string keyword, int page, int pageSize)
         {
             var model = _billService.GetAllPaging(startDate, endDate, keyword, page, pageSize);
             return new OkObjectResult(model);
         }
+
         [HttpPost]
         public IActionResult SaveEntity(BillViewModel billVm)
         {
@@ -70,6 +74,7 @@ namespace TeduCoreApp.Areas.Admin.Controllers
             _billService.Save();
             return new OkObjectResult(billVm);
         }
+
         [HttpGet]
         public IActionResult GetPaymentMethod()
         {
@@ -107,6 +112,7 @@ namespace TeduCoreApp.Areas.Admin.Controllers
             var sizes = _billService.GetSizes();
             return new OkObjectResult(sizes);
         }
+
         [HttpPost]
         public IActionResult ExportExcel(int billId)
         {
@@ -163,7 +169,6 @@ namespace TeduCoreApp.Areas.Admin.Controllers
                     worksheet.Cells[26, 1].Value = numberWord;
                     var billDate = billDetail.DateCreated;
                     worksheet.Cells[28, 3].Value = billDate.Day + ", " + billDate.Month + ", " + billDate.Year;
-
 
                     package.SaveAs(file); //Save the workbook.
                 }
