@@ -15,17 +15,19 @@ namespace TeduCoreApp.Application.Implementation
     {
         private IRepository<Contact, string> _contactRepository;
         private IUnitOfWork _unitOfWork;
+        private readonly IMapper _mapper;
 
         public ContactService(IRepository<Contact, string> contactRepository,
-            IUnitOfWork unitOfWork)
+            IUnitOfWork unitOfWork, IMapper mapper)
         {
             this._contactRepository = contactRepository;
             this._unitOfWork = unitOfWork;
+            _mapper = mapper;
         }
 
         public void Add(ContactViewModel pageVm)
         {
-            var page = Mapper.Map<ContactViewModel, Contact>(pageVm);
+            var page = _mapper.Map<ContactViewModel, Contact>(pageVm);
             _contactRepository.Add(page);
         }
 
@@ -41,7 +43,8 @@ namespace TeduCoreApp.Application.Implementation
 
         public List<ContactViewModel> GetAll()
         {
-            return _contactRepository.FindAll().ProjectTo<ContactViewModel>().ToList();
+            return _mapper.ProjectTo<ContactViewModel>(
+                _contactRepository.FindAll()).ToList();
         }
 
         public PagedResult<ContactViewModel> GetAllPaging(string keyword, int page, int pageSize)
@@ -57,7 +60,7 @@ namespace TeduCoreApp.Application.Implementation
 
             var paginationSet = new PagedResult<ContactViewModel>()
             {
-                Results = data.ProjectTo<ContactViewModel>().ToList(),
+                Results = _mapper.ProjectTo<ContactViewModel>(data).ToList(),
                 CurrentPage = page,
                 RowCount = totalRow,
                 PageSize = pageSize
@@ -68,7 +71,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public ContactViewModel GetById(string id)
         {
-            return Mapper.Map<Contact, ContactViewModel>(_contactRepository.FindById(id));
+            return _mapper.Map<Contact, ContactViewModel>(_contactRepository.FindById(id));
         }
 
         public void SaveChanges()
@@ -78,7 +81,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public void Update(ContactViewModel pageVm)
         {
-            var page = Mapper.Map<ContactViewModel, Contact>(pageVm);
+            var page = _mapper.Map<ContactViewModel, Contact>(pageVm);
             _contactRepository.Update(page);
         }
     }
