@@ -15,17 +15,19 @@ namespace TeduCoreApp.Application.Implementation
     {
         private IRepository<Feedback, int> _feedbackRepository;
         private IUnitOfWork _unitOfWork;
+        private readonly IMapper _mapper;
 
         public FeedbackService(IRepository<Feedback, int> feedbackRepository,
-            IUnitOfWork unitOfWork)
+            IUnitOfWork unitOfWork, IMapper mapper)
         {
             _feedbackRepository = feedbackRepository;
             _unitOfWork = unitOfWork;
+            _mapper = mapper;
         }
 
         public void Add(FeedbackViewModel feedbackVm)
         {
-            var page = Mapper.Map<FeedbackViewModel, Feedback>(feedbackVm);
+            var page = _mapper.Map<FeedbackViewModel, Feedback>(feedbackVm);
             _feedbackRepository.Add(page);
         }
 
@@ -41,7 +43,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public List<FeedbackViewModel> GetAll()
         {
-            return _feedbackRepository.FindAll().ProjectTo<FeedbackViewModel>().ToList();
+            return _mapper.ProjectTo<FeedbackViewModel>(_feedbackRepository.FindAll()).ToList();
         }
 
         public PagedResult<FeedbackViewModel> GetAllPaging(string keyword, int page, int pageSize)
@@ -57,7 +59,7 @@ namespace TeduCoreApp.Application.Implementation
 
             var paginationSet = new PagedResult<FeedbackViewModel>()
             {
-                Results = data.ProjectTo<FeedbackViewModel>().ToList(),
+                Results = _mapper.ProjectTo<FeedbackViewModel>(data).ToList(),
                 CurrentPage = page,
                 RowCount = totalRow,
                 PageSize = pageSize
@@ -68,7 +70,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public FeedbackViewModel GetById(int id)
         {
-            return Mapper.Map<Feedback, FeedbackViewModel>(_feedbackRepository.FindById(id));
+            return _mapper.Map<Feedback, FeedbackViewModel>(_feedbackRepository.FindById(id));
         }
 
         public void SaveChanges()
@@ -78,7 +80,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public void Update(FeedbackViewModel feedbackVm)
         {
-            var page = Mapper.Map<FeedbackViewModel, Feedback>(feedbackVm);
+            var page = _mapper.Map<FeedbackViewModel, Feedback>(feedbackVm);
             _feedbackRepository.Update(page);
         }
     }
