@@ -25,7 +25,7 @@ namespace TeduCoreApp.Application.Implementation
         private IRepository<ProductQuantity, int> _productQuantityRepository;
         private IRepository<ProductImage, int> _productImageRepository;
         private IRepository<WholePrice, int> _wholePriceRepository;
-
+        private readonly IMapper _mapper;
         private IUnitOfWork _unitOfWork;
 
         public ProductService(IRepository<Product, int> productRepository,
@@ -34,7 +34,7 @@ namespace TeduCoreApp.Application.Implementation
             IRepository<ProductImage, int> productImageRepository,
             IRepository<WholePrice, int> wholePriceRepository,
         IUnitOfWork unitOfWork,
-        IRepository<ProductTag, int> productTagRepository)
+        IRepository<ProductTag, int> productTagRepository, IMapper mapper)
         {
             _productRepository = productRepository;
             _tagRepository = tagRepository;
@@ -43,6 +43,7 @@ namespace TeduCoreApp.Application.Implementation
             _wholePriceRepository = wholePriceRepository;
             _productImageRepository = productImageRepository;
             _unitOfWork = unitOfWork;
+            _mapper = mapper;
         }
 
         public ProductViewModel Add(ProductViewModel productVm)
@@ -71,7 +72,7 @@ namespace TeduCoreApp.Application.Implementation
                     };
                     productTags.Add(productTag);
                 }
-                var product = Mapper.Map<ProductViewModel, Product>(productVm);
+                var product = _mapper.Map<ProductViewModel, Product>(productVm);
                 foreach (var productTag in productTags)
                 {
                     product.ProductTags.Add(productTag);
@@ -108,7 +109,9 @@ namespace TeduCoreApp.Application.Implementation
 
         public List<ProductViewModel> GetAll()
         {
-            return _productRepository.FindAll(x => x.ProductCategory).ProjectTo<ProductViewModel>().ToList();
+            return _mapper.ProjectTo<ProductViewModel>(
+                _productRepository.FindAll(x => x.ProductCategory))
+                .ToList();
         }
 
         public PagedResult<ProductViewModel> GetAllPaging(int? categoryId, string keyword, int page, int pageSize)
@@ -124,7 +127,7 @@ namespace TeduCoreApp.Application.Implementation
             query = query.OrderByDescending(x => x.DateCreated)
                 .Skip((page - 1) * pageSize).Take(pageSize);
 
-            var data = query.ProjectTo<ProductViewModel>().ToList();
+            var data = _mapper.ProjectTo<ProductViewModel>(query).ToList();
 
             var paginationSet = new PagedResult<ProductViewModel>()
             {
@@ -138,12 +141,14 @@ namespace TeduCoreApp.Application.Implementation
 
         public ProductViewModel GetById(int id)
         {
-            return Mapper.Map<Product, ProductViewModel>(_productRepository.FindById(id));
+            return _mapper.Map<Product, ProductViewModel>(_productRepository.FindById(id));
         }
 
         public List<ProductQuantityViewModel> GetQuantities(int productId)
         {
-            return _productQuantityRepository.FindAll(x => x.ProductId == productId).ProjectTo<ProductQuantityViewModel>().ToList();
+            return _mapper.ProjectTo<ProductQuantityViewModel>(
+                _productQuantityRepository.FindAll(x => x.ProductId == productId))
+                .ToList();
         }
 
         public void ImportExcel(string filePath, int categoryId)
@@ -218,7 +223,7 @@ namespace TeduCoreApp.Application.Implementation
                 }
             }
 
-            var product = Mapper.Map<ProductViewModel, Product>(productVm);
+            var product = _mapper.Map<ProductViewModel, Product>(productVm);
             foreach (var productTag in productTags)
             {
                 product.ProductTags.Add(productTag);
@@ -228,8 +233,9 @@ namespace TeduCoreApp.Application.Implementation
 
         public List<ProductImageViewModel> GetImages(int productId)
         {
-            return _productImageRepository.FindAll(x => x.ProductId == productId)
-                .ProjectTo<ProductImageViewModel>().ToList();
+            return _mapper.ProjectTo<ProductImageViewModel>(
+                _productImageRepository.FindAll(x => x.ProductId == productId)
+                ).ToList();
         }
 
         public void AddImages(int productId, string[] images)
@@ -263,41 +269,47 @@ namespace TeduCoreApp.Application.Implementation
 
         public List<WholePriceViewModel> GetWholePrices(int productId)
         {
-            return _wholePriceRepository.FindAll(x => x.ProductId == productId).ProjectTo<WholePriceViewModel>().ToList();
+            return _mapper.ProjectTo<WholePriceViewModel>(
+                _wholePriceRepository.FindAll(x => x.ProductId == productId)
+                ).ToList();
         }
 
         public List<ProductViewModel> GetLastest(int top)
         {
-            return _productRepository.FindAll(x => x.Status == Status.Active).OrderByDescending(x => x.DateCreated)
-                .Take(top).ProjectTo<ProductViewModel>().ToList();
+            return _mapper.ProjectTo<ProductViewModel>(
+                _productRepository.FindAll(x => x.Status == Status.Active)
+                .OrderByDescending(x => x.DateCreated)
+                )
+                .ToList();
         }
 
         public List<ProductViewModel> GetHotProduct(int top)
         {
-            return _productRepository.FindAll(x => x.Status == Status.Active && x.HotFlag == true)
+            return _mapper.ProjectTo<ProductViewModel>(
+                _productRepository.FindAll(x => x.Status == Status.Active && x.HotFlag == true)
                 .OrderByDescending(x => x.DateCreated)
-                .Take(top)
-                .ProjectTo<ProductViewModel>()
+                .Take(top))
+
                 .ToList();
         }
 
         public List<ProductViewModel> GetRelatedProducts(int id, int top)
         {
             var product = _productRepository.FindById(id);
-            return _productRepository.FindAll(x => x.Status == Status.Active
+            return _mapper.ProjectTo<ProductViewModel>(
+                _productRepository.FindAll(x => x.Status == Status.Active
                 && x.Id != id && x.CategoryId == product.CategoryId)
             .OrderByDescending(x => x.DateCreated)
-            .Take(top)
-            .ProjectTo<ProductViewModel>()
+            .Take(top))
             .ToList();
         }
 
         public List<ProductViewModel> GetUpsellProducts(int top)
         {
-            return _productRepository.FindAll(x => x.PromotionPrice != null)
+            return _mapper.ProjectTo<ProductViewModel>(
+                _productRepository.FindAll(x => x.PromotionPrice != null)
                .OrderByDescending(x => x.DateModified)
-               .Take(top)
-               .ProjectTo<ProductViewModel>().ToList();
+               .Take(top)).ToList();
         }
 
         public List<TagViewModel> GetProductTags(int productId)
