@@ -21,7 +21,7 @@ namespace TeduCoreApp.Application.Implementation
         private IRepository<Permission, int> _permissionRepository;
         private IRepository<Announcement, string> _announRepository;
         private IRepository<AnnouncementUser, int> _announUserRepository;
-
+        private IMapper _mapper;
         private IUnitOfWork _unitOfWork;
 
         public RoleService(RoleManager<AppRole> roleManager,
@@ -29,7 +29,7 @@ namespace TeduCoreApp.Application.Implementation
             IRepository<AnnouncementUser, int> announUserRepository,
          IRepository<Function, string> functionRepository,
          IRepository<Permission, int> permissionRepository,
-            IRepository<Announcement, string> announRepository)
+            IRepository<Announcement, string> announRepository, IMapper mapper)
         {
             _unitOfWork = unitOfWork;
             _roleManager = roleManager;
@@ -37,6 +37,7 @@ namespace TeduCoreApp.Application.Implementation
             _functionRepository = functionRepository;
             _announUserRepository = announUserRepository;
             _permissionRepository = permissionRepository;
+            _mapper = mapper;
         }
 
         public async Task<bool> AddAsync(AnnouncementViewModel announcementVm,
@@ -48,11 +49,11 @@ namespace TeduCoreApp.Application.Implementation
                 Description = roleVm.Description
             };
             var result = await _roleManager.CreateAsync(role);
-            var announcement = Mapper.Map<AnnouncementViewModel, Announcement>(announcementVm);
+            var announcement = _mapper.Map<AnnouncementViewModel, Announcement>(announcementVm);
             _announRepository.Add(announcement);
             foreach (var userVm in announcementUsers)
             {
-                var user = Mapper.Map<AnnouncementUserViewModel, AnnouncementUser>(userVm);
+                var user = _mapper.Map<AnnouncementUserViewModel, AnnouncementUser>(userVm);
                 _announUserRepository.Add(user);
             }
             _unitOfWork.Commit();
@@ -83,7 +84,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public async Task<List<AppRoleViewModel>> GetAllAsync()
         {
-            return await _roleManager.Roles.ProjectTo<AppRoleViewModel>().ToListAsync();
+            return await _mapper.ProjectTo<AppRoleViewModel>(_roleManager.Roles).ToListAsync();
         }
 
         public PagedResult<AppRoleViewModel> GetAllPagingAsync(string keyword, int page, int pageSize)
@@ -97,7 +98,7 @@ namespace TeduCoreApp.Application.Implementation
             query = query.Skip((page - 1) * pageSize)
                .Take(pageSize);
 
-            var data = query.ProjectTo<AppRoleViewModel>().ToList();
+            var data = _mapper.ProjectTo<AppRoleViewModel>(query).ToList();
             var paginationSet = new PagedResult<AppRoleViewModel>()
             {
                 Results = data,
@@ -112,7 +113,7 @@ namespace TeduCoreApp.Application.Implementation
         public async Task<AppRoleViewModel> GetById(Guid id)
         {
             var role = await _roleManager.FindByIdAsync(id.ToString());
-            return Mapper.Map<AppRole, AppRoleViewModel>(role);
+            return _mapper.Map<AppRole, AppRoleViewModel>(role);
         }
 
         public List<PermissionViewModel> GetListFunctionWithRole(Guid roleId)
@@ -138,7 +139,7 @@ namespace TeduCoreApp.Application.Implementation
 
         public void SavePermission(List<PermissionViewModel> permissionVms, Guid roleId)
         {
-            var permissions = Mapper.Map<List<PermissionViewModel>, List<Permission>>(permissionVms);
+            var permissions = _mapper.Map<List<PermissionViewModel>, List<Permission>>(permissionVms);
             var oldPermission = _permissionRepository.FindAll().Where(x => x.RoleId == roleId).ToList();
             if (oldPermission.Count > 0)
             {
