@@ -1,5 +1,4 @@
-﻿
-@{
+﻿@{
     ViewData["Title"] = "Index";
 }
 @section Styles{
@@ -24,7 +23,6 @@
         <div class="col-md-2 col-sm-2 col-xs-12 form-group pull-right">
             <button class="btn btn-success" id="btnCreate" type="button">Create</button>
         </div>
-
     </div>
 </div>
 <div class="clearfix"></div>
@@ -42,4 +40,4 @@
         </div>
     </div>
 </div>
-@Html.Partial("_AddEditModal.cshtml")
+<partial name="_AddEditModal" />
