@@ -39,9 +39,7 @@
                         <button class="btn btn-default" id="btn-search" type="button">Search!</button>
                     </span>
                 </div>
-
             </div>
-
         </div>
     </div>
 
@@ -86,12 +84,10 @@
                             </div>
                         </div>
                     </div>
-
                 </div>
             </div>
         </div>
     </div>
-
 
     <script id="table-template" type="x-tmpl-mustache">
         <tr>
@@ -115,7 +111,6 @@
             </td>
         </tr>
     </script>
-
 </div>
 <script id="role-template" type="x-tmpl-mustache">
     <div class="checkbox">
@@ -125,4 +120,4 @@
         </label>
     </div>
 </script>
-@Html.Partial("_AddEditModal.cshtml")
+<partial name="_AddEditModal" />
