@@ -5,7 +5,7 @@
 }
 
 <h4>Set your password</h4>
-@Html.Partial("_StatusMessage", Model.StatusMessage)
+<partial name="_StatusMessage" model="Model.StatusMessage" />
 <p class="text-info">
     You do not have a local username/password for this site. Add a local
     account so you can log in without an external login.
